Midpoint + Small fixes
I added a Midpoint rule integrator. I did this using the midpoint Riemann method (where the evaluation point is the center of the node... this is not the node-centric midpoint rule). I decided to do this using the Reassociation transformation (rather than multiple accumulators). I have no idea which one is better in this case, but I wanted to try both.

As expected, the midpoint method is far far more accurate (when compared to data from Mathematica) than the Left or Right Riemann method.

I also caught a small bug in the Right Riemann code (see Riemann.cc:106).