Added Trapezoidal + Fixed a bug in Riemann
I added code for the Trapezoidal method. Unsurprisingly, this method is extremely accurate (just about as accurate as the midpoint method).

I decided to implement the Trapezoidal method using multiple accumulators. I plan on actually performing a "Cycles per element" test of this code in the near future. This will hopefully allow me to determine how good my optimizations actually are.

I also fixed a bug in the Reimann code. I the main loop, I compared i to n, the number of subdomains, minus 3. The issue here is that n is an unsigned int. Therefore, C was calculating n -3 as an unsigned integer. If n is < 3, then n-3, as an unsigned integer, overflows. This lead to a faulty test condition (and a segementation fault!). To fix this, I typecasted n to an integer (signed) before subtracting 3 (and then comparing to i).

I also made a changes (mostly syntax) to the Riemann code.

All 4 integral approximations are running!