Basic derivative approximations
I added code for the Forward/Backward two point finite difference approximations. I also added code for the central difference approximation (which is a 3 point method). I plan on adding code for 3 and 5 point approximations.

The new derivative approximations are in a new namespace, the Derivative namespace.

I also added a compile guard to Trapezoidal.